requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backcall==0.2.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.5
cryptography==38.0.1
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
filelock==3.4.2
flit_core==3.6.0
frozenlist==1.3.3
fsspec==2021.5.0
future==0.18.2
gast==0.3.3
google-auth==2.6.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.42.0
h5py==2.10.0
huggingface-hub==0.2.1
humanize==3.13.1
idna==3.4
imageio==2.22.4
importlib-metadata==4.11.3
ipykernel==6.15.2
ipython==7.31.1
jedi==0.18.1
joblib==1.0.1
jupyter_client==7.4.7
jupyter_core==4.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monai==0.8.0
multidict==6.0.2
nest-asyncio==1.5.5
nibabel==3.2.1
numexpr==2.8.4
numpy==1.21.5
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pip==22.2.2
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.0
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.3.6
pytz==2022.1
pywin32==305.1
PyYAML==5.4.1
pyzmq==23.2.0
regex==2021.11.10
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.7.3
setuptools==65.5.0
SimpleITK==2.1.1
six==1.16.0
sklearn==0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.6.0
termcolor==2.1.0
threadpoolctl==2.1.0
tifffile==2021.11.2
tokenizers==0.10.3
torch==1.9.0
torchaudio==0.9.0
torchio==0.18.71
torchmetrics==0.3.2
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.2
tqdm==4.61.1
traitlets==5.1.1
transformers==4.15.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.0
